Print movelist. Add pawn movegen.
