--- .vscode/launch.json
@@ -1,45 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug executable 'vault'",
-            "cargo": {
-                "args": [
-                    "build",
-                    "--bin=vault",
-                    "--package=vault"
-                ],
-                "filter": {
-                    "name": "vault",
-                    "kind": "bin"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        },
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug unit tests in executable 'vault'",
-            "cargo": {
-                "args": [
-                    "test",
-                    "--no-run",
-                    "--bin=vault",
-                    "--package=vault"
-                ],
-                "filter": {
-                    "name": "vault",
-                    "kind": "bin"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        }
-    ]
-}
