+++ src/movegen/mod.rs
@@ -1,6 +1,7 @@
 pub mod attacks;
 pub mod generate;
 
+use core::fmt;
 use std::sync::LazyLock;
 
 use crate::{board::Board, defs::{Color, Piece, PieceType, Ranks, RANKS_BOARD}};
@@ -38,17 +39,17 @@ impl MoveList {
 
     fn set_score(mv: u32, score: u64) -> u64 { return (mv as u64) | (score << 47) } 
 
-    pub fn add_quiet_move(&mut self, position: &Board, mv: u32) {
+    fn add_quiet_move(&mut self, position: &Board, mv: u32) {
         // TODO set score killer move
         self.add(MoveList::set_score(mv, 0));
     }
 
-    pub fn add_capture_move(&mut self, position: &Board, mv: u32) {
+    fn add_capture_move(&mut self, position: &Board, mv: u32) {
         // TODO set score MVV_LVA
         self.add(MoveList::set_score(mv, 10000));
     }
 
-    pub fn add_quiet_pawn_move(&mut self, position: &Board, from: usize, to: usize) {
+    fn add_quiet_pawn_move(&mut self, position: &Board, from: usize, to: usize) {
         let (promotion_rank, queen, rook, bishop, knight) = if position.side == Color::White {
             (
                 Ranks::Rank7, 
@@ -107,7 +108,7 @@ impl MoveList {
         }
     }
 
-    pub fn add_capture_pawn_move(&mut self, position: &Board, from: usize, to: usize, captured: usize) {
+    fn add_capture_pawn_move(&mut self, position: &Board, from: usize, to: usize, captured: usize) {
         let (promotion_rank, queen, rook, bishop, knight) = if position.side == Color::White {
             (
                 Ranks::Rank7, 
@@ -166,13 +167,195 @@ impl MoveList {
         }
     }
 
-    pub fn add_en_passant_move(&mut self, position: &Board, from: usize, to: usize, captured: usize) {
-        
+    pub fn generate_pawn_moves(&mut self, position: Board) {
+        let occupied_bb = [position.occupancies(Color::White), position.occupancies(Color::Black)];
+        let empty_squares: u64 = !(occupied_bb[Color::White as usize] | occupied_bb[Color::Black as usize]);
+        let color = position.side;
+        let (
+            our_pawns,
+            their_pieces,
+            push_offset,
+            double_push_rank,
+            pawn_attacks,
+            en_passant_attackers_table
+        ) = if color == Color::White {
+            (
+                position.bitboards[PieceType::Pawn.bb_index(Color::White)],
+                occupied_bb[Color::Black as usize],
+                8,
+                RANK_BB_MASK[3],
+                &WHITE_PAWN_ATTACKS,
+                &BLACK_PAWN_ATTACKS
+            )
+        } else {
+            (
+                position.bitboards[PieceType::Pawn.bb_index(Color::Black)],
+                occupied_bb[Color::White as usize],
+                -8,
+                RANK_BB_MASK[6],
+                &BLACK_PAWN_ATTACKS,
+                &WHITE_PAWN_ATTACKS
+            )
+        };
+
+        let mut single_pushes = if color == Color::White {
+            our_pawns.wrapping_shl(push_offset as u32) & empty_squares
+        } else {
+            our_pawns.wrapping_shr((-push_offset) as u32) & empty_squares
+        };
+
+        let mut double_pushes = if color == Color::White {
+            (single_pushes & double_push_rank).wrapping_shl(push_offset as u32) & empty_squares
+        } else {
+            (single_pushes & double_push_rank).wrapping_shr((-push_offset) as u32) & empty_squares
+        };
+
+        while single_pushes != 0 {
+            let to_square = single_pushes.trailing_zeros() as usize;
+            let from_square = (to_square as i32 - push_offset) as usize;
+            self.add_quiet_pawn_move(&position, from_square, to_square);
+
+            single_pushes &= single_pushes - 1;
+        }
+
+        while double_pushes != 0 {
+            let to_square = double_pushes.trailing_zeros() as usize;
+            let from_square = (to_square as i32 - (push_offset * 2)) as usize;
+            self.add_quiet_move(&position, MoveList::move_builder(
+                from_square, 
+                to_square, 
+                0, 
+                0 /*Pawn start*/, 
+                0
+            ));
+
+            double_pushes &= double_pushes - 1;
+        }
+
+        let mut pawns = our_pawns;
+        while pawns != 0 {
+            let from_square = pawns.trailing_zeros() as usize;
+            let mut valid_captures = pawn_attacks[from_square] & their_pieces;
+
+            while valid_captures != 0 {
+                let to_square = valid_captures.trailing_zeros() as usize;
+                let captured_piece = position.pieces[to_square];
+
+                if let Some(piece) = captured_piece {
+                    self.add_capture_pawn_move(&position, from_square, to_square, piece.bb_index());
+                }
+
+                valid_captures &= valid_captures - 1;
+            }
+            pawns &= pawns - 1;
+        }
+
+        if let Some(en_passant_square) = position.en_passant {
+            let potential_attackers = en_passant_attackers_table[en_passant_square];
+            let mut en_passant_attackers = potential_attackers & our_pawns;
+
+            while en_passant_attackers != 0 {
+                let from_square = en_passant_attackers.trailing_zeros() as usize;
+                if let Some(piece) = position.pieces[from_square] {
+                    self.add_capture_pawn_move(&position, from_square, en_passant_square, piece.bb_index());
+                }
+
+                en_passant_attackers &= en_passant_attackers - 1;
+            }
+        }
     }
 
     pub fn len(&self) -> usize { self.count }
 
     pub fn iter(&self) -> std::slice::Iter<'_, u64> { self.moves[..self.count].iter() }
+}
+
+const SQUARE_TO_STRING: [&str; 64] = [
+    "a1", "b1", "c1", "d1", "e1", "f1", "g1", "h1",
+    "a2", "b2", "c2", "d2", "e2", "f2", "g2", "h2",
+    "a3", "b3", "c3", "d3", "e3", "f3", "g3", "h3",
+    "a4", "b4", "c4", "d4", "e4", "f4", "g4", "h4",
+    "a5", "b5", "c5", "d5", "e5", "f5", "g5", "h5",
+    "a6", "b6", "c6", "d6", "e6", "f6", "g6", "h6",
+    "a7", "b7", "c7", "d7", "e7", "f7", "g7", "h7",
+    "a8", "b8", "c8", "d8", "e8", "f8", "g8", "h8",
+];
+
+/// Returns the promotion piece character based on the piece index.
+///
+/// This function makes assumptions based on your `add_..._pawn_move` functions:
+/// 1. `promote: 0` is used for no-promotion.
+/// 2. The piece indices for Q, R, B, N are passed for promotion.
+/// 3. We assume a common encoding like:
+///    W_Q=5, W_R=4, W_B=3, W_N=2
+///    B_Q=11, B_R=10, B_B=9, B_N=8
+///    (This order matches your `queen`, `rook`, `bishop`, `knight` calls)
+fn get_promo_char(promo_index: u32) -> &'static str {
+    match promo_index {
+        // White pieces (assuming N=2, B=3, R=4, Q=5)
+        2 => "n",
+        3 => "b",
+        4 => "r",
+        5 => "q",
+        // Black pieces (assuming n=8, b=9, r=10, q=11)
+        8 => "n",
+        9 => "b",
+        10 => "r",
+        11 => "q",
+        // Default: no promotion (index 0) or invalid (King/Pawn)
+        _ => "",
+    }
+}
+
+impl fmt::Display for MoveList {
+    /// Formats the move list for printing.
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        // Write a header
+        writeln!(f, "Move List ({} moves found):", self.count)?;
+        
+        // Iterate over all moves in the list
+        for i in 0..self.count {
+            // Get the 64-bit move entry
+            let move_with_score = self.moves[i];
+
+            // Extract the 32-bit move data (lower 32 bits)
+            let move_data = move_with_score as u32;
+
+            // Extract the score (upper bits)
+            let score = move_with_score >> 47;
+
+            // --- Decode the 32-bit move data ---
+            // Based on: move_builder(from, to, capture, flags, promote)
+            
+            // from: bits 0-5 (mask 0x3F)
+            let from = (move_data & 0x3F) as usize;
+            
+            // to: bits 6-11 (mask 0x3F)
+            let to = ((move_data >> 6) & 0x3F) as usize;
+            
+            // promote: bits 20-24 (mask 0x1F)
+            let promote = (move_data >> 20) & 0x1F;
+
+            // Convert square indices to algebraic notation
+            let from_sq_str = SQUARE_TO_STRING[from];
+            let to_sq_str = SQUARE_TO_STRING[to];
+            
+            // Get the promotion character (e.g., "q", "r", or "")
+            let promo_str = get_promo_char(promote as u32);
+
+            // Write the formatted move string
+            // Example: "  1: e2e4 (Score: 0)"
+            // Example: " 12: e7e8q (Score: 10000)"
+            writeln!(f, "  {:>2}: {}{}{} (Score: {})",
+                     i + 1,
+                     from_sq_str,
+                     to_sq_str,
+                     promo_str,
+                     score)?;
+        }
+        
+        Ok(())
+    }
 }
 
 /*
